Refactor service class and move to new file
